Change to byte for hashcode

+++ src/main/java/net/minestom/server/collision/CollisionUtils.java
@@ -184,7 +184,7 @@ public final class CollisionUtils {
 
     @ApiStatus.Internal
     public static Shape parseBlockShape(Map<Object, Object> internCache, String collision, String occlusion, boolean occludes, byte lightEmission) {
-        record ShapeEntry(String collision, String occlusion, boolean occludes, int lightEmission) {} // Easy way to Hashcode
+        record ShapeEntry(String collision, String occlusion, boolean occludes, byte lightEmission) {} // Easy way to Hashcode
         ShapeEntry entry = new ShapeEntry(collision, occlusion, occludes, lightEmission);
         final Shape cachedShape = (Shape) internCache.get(entry);
         if (cachedShape != null) return cachedShape;
